(-_-)
Welp, it is now even more nixed, and I hate that it is so funky like
that you know. I wish there were some more simple functions, such as
exponent but it makes sense that they don't exist. But the fact that you
cannot convert a Attribute set to a higher level is so annoying. Or that
there is no good union of attribute sets.